refactor: export * from select attribute component to apikey dialog
