+++ src/frontend/components/administration/apikey/CreateDialog.vue
@@ -38,7 +38,6 @@
             />
           </v-card-actions>
           <SettingsAttributes
-            :all="attributesAll"
             :simple="attributesSimple"
             :best-oa-location="attributesBestOaLocation"
             :first-oa-location="attributesFirstOaLocation"
@@ -71,7 +70,7 @@
 </template>
 
 <script>
-import SettingsAttributes from '~/components/administration/SettingsAttributes.vue'
+import SettingsAttributes from '~/components/unpaywallArgs/SettingsAttributes.vue'
 
 export default {
   name: 'ApikeyCreateDialog',
@@ -122,9 +121,6 @@ export default {
     },
     attributesZAuthors () {
       return this.attributes?.filter(e => e.includes('z_authors')).map(e => e.split('.')[1])
-    },
-    attributesAll () {
-      return this.attributes.includes('*')
     }
   },
   methods: {
@@ -158,7 +154,11 @@ export default {
       this.updateVisible(false)
     },
     updateAttributes (attributesSelected) {
-      this.attributes = attributesSelected
+      if (attributesSelected.length === 50) {
+        this.attributes = ['*']
+      } else {
+        this.attributes = attributesSelected
+      }
     }
   }
 }
